+++ api/tests/team/integration/domain/usecases/find-user-organizations-for-admin.usecase.test.js
@@ -1,4 +1,4 @@
-import { UserOrganizationForAdmin } from '../../../../../src/shared/domain/read-models/UserOrganizationForAdmin.js';
+import { UserOrganizationForAdmin } from '../../../../../src/team/domain/read-models/UserOrganizationForAdmin.js';
 import { usecases } from '../../../../../src/team/domain/usecases/index.js';
 import { databaseBuilder, expect } from '../../../../test-helper.js';
